+++ docs/src/components/page-shell.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import PageLayout from '../../../src/components/page-layout';
-import navigation from '@mapbox/batfish/data/navigation'; // eslint-disable-line
-import topics from '@mapbox/batfish/data/topics'; // eslint-disable-line
+import navigation from '@mapbox/batfish/data/navigation';
+import filters from '@mapbox/batfish/data/filters';
 import constants from '../constants';
-
+import AppropriateImage from './appropriate-image';
 import categories from '../categories.json';
 
 const slug = (string) => string.toLowerCase();
@@ -32,8 +32,7 @@ class PageShell extends React.Component {
 
     return (
       <PageLayout
-        includeFilterBar={true}
-        topics={topics}
+        filters={filters}
         headings={headings}
         frontMatter={{
           ...frontMatter,
@@ -50,6 +49,7 @@ class PageShell extends React.Component {
         location={location}
         constants={constants}
         navigation={navigation}
+        AppropriateImage={AppropriateImage}
       >
         {children}
       </PageLayout>