+++ src/components/page-layout/components/content.js
@@ -1,23 +1,50 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import ExampleIndex from './example-index.js';
 import classnames from 'classnames';
-import LayoutExamples from './layouts/example/content.js';
 import OnThisPage from '../../on-this-page/on-this-page';
 import Feedback from '../../feedback/feedback';
 import OverviewHeader from '../../overview-header/overview-header';
 
 export default class Content extends React.Component {
-  renderExamplesIndex = () => {
-    const { frontMatter, parentPath, topics, AppropriateImage } = this.props;
-    return (
-      <LayoutExamples
-        topics={topics[parentPath].topics}
+  render() {
+    const {
+      children,
+      layoutConfig,
+      filters,
+      parentPath,
+      frontMatter,
+      AppropriateImage
+    } = this.props;
+    const { showCards } = layoutConfig;
+    // if showCards is true, use the ExampleIndex component
+    // else use the default content wrapper to display the content
+    return showCards ? (
+      <ExampleIndex
+        {...this.props}
+        filters={filters[parentPath]}
         frontMatter={frontMatter}
         sectionPath={parentPath}
         AppropriateImage={AppropriateImage}
       />
+    ) : (
+      <ContentWrapper {...this.props}>{children}</ContentWrapper>
     );
-  };
+  }
+}
+
+Content.propTypes = {
+  children: PropTypes.node,
+  frontMatter: PropTypes.object.isRequired,
+  parentPath: PropTypes.string,
+  filters: PropTypes.object,
+  AppropriateImage: PropTypes.func,
+  layoutConfig: PropTypes.shape({
+    showCards: PropTypes.bool
+  })
+};
+
+export class ContentWrapper extends React.Component {
   renderFeedback = () => {
     const { location, section, layoutConfig } = this.props;
     const { layout } = layoutConfig;
@@ -35,14 +62,11 @@ export default class Content extends React.Component {
     );
   };
 
-  renderAside = () => {
+  renderAside = (showFeedback) => {
     const { frontMatter } = this.props;
-    const { hideFeedback, layout } = frontMatter;
+    const { layout } = frontMatter;
     const headings = frontMatter.headings || this.props.headings;
-    const showToc =
-      headings &&
-      headings.length > 0 &&
-      (layout === 'page' || layout === 'accordion');
+    const showToc = headings && headings.length > 0 && layout === 'page';
     return (
       <aside
         data-swiftype-index="false"
@@ -53,7 +77,7 @@ export default class Content extends React.Component {
         {showToc && headings && headings.length > 0 && (
           <OnThisPage headings={headings} themeWrapper="mb36-mxl mb18" />
         )}
-        {!hideFeedback && (
+        {showFeedback && (
           <div className="none block-mxl">{this.renderFeedback()}</div>
         )}
       </aside>
@@ -69,16 +93,15 @@ export default class Content extends React.Component {
       layout,
       overviewHeader
     } = frontMatter;
-    const { hideTitle, showCards, sidebar } = layoutConfig;
+    const { hideTitle } = layoutConfig;
+
+    // check frontmatter then default to layout config
+    const showFeedback = hideFeedback
+      ? !hideFeedback
+      : !layoutConfig.hideFeedback;
+
     return (
-      <div
-        id="docs-content"
-        className={classnames('pr0-mm mt24', {
-          'mb60 px24-mm': sidebar !== 'none',
-          'px24-mm': sidebar === 'none',
-          'mt24-mm mt60 pt30 pt0-mm': sidebar === 'accordion' // clear the mobile sticky nav
-        })}
-      >
+      <div id="docs-content">
         {!hideTitle && (
           <div
             className={classnames('col prose', {
@@ -92,8 +115,8 @@ export default class Content extends React.Component {
 
         <div className="grid grid--gut60">
           {(layoutConfig.aside !== 'none' || this.props.customAside) && (
-            <div className="dr-ui--page-layout-aside col col--4-mxl">
-              {this.renderAside()}
+            <div className="dr-ui--page-layout-aside col col--12 col--4-mxl">
+              {this.renderAside(showFeedback)}
             </div>
           )}
 
@@ -104,8 +127,7 @@ export default class Content extends React.Component {
             })}
           >
             {children}
-            {showCards ? this.renderExamplesIndex() : ''}
-            {!hideFeedback && (
+            {showFeedback && (
               <div
                 className={classnames('my36', {
                   'block none-mxl': layout !== 'example' && layout !== 'full' // hide feedback at bottom of page on larger screens, unless layout is example or full (always show it on the bottom)
@@ -121,7 +143,7 @@ export default class Content extends React.Component {
   }
 }
 
-Content.propTypes = {
+ContentWrapper.propTypes = {
   children: PropTypes.node,
   frontMatter: PropTypes.shape({
     title: PropTypes.string.isRequired,
@@ -134,7 +156,7 @@ Content.propTypes = {
   headings: PropTypes.array,
   location: PropTypes.object.isRequired,
   parentPath: PropTypes.string,
-  topics: PropTypes.object,
+  filters: PropTypes.object,
   AppropriateImage: PropTypes.func,
   section: PropTypes.string,
   constants: PropTypes.shape({
@@ -150,8 +172,7 @@ Content.propTypes = {
     hideTitle: PropTypes.bool,
     showCards: PropTypes.bool,
     hideFeedback: PropTypes.bool,
-    aside: PropTypes.string,
-    sidebar: PropTypes.string
+    aside: PropTypes.string
   }),
   navigation: PropTypes.shape({
     hierarchy: PropTypes.object