+++ src/components/page-layout/components/sidebar.js
@@ -2,32 +2,25 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import NavigationAccordion from '../../navigation-accordion';
 import ProductMenu from '../../product-menu/product-menu';
+import classnames from 'classnames';
 
 export default class Sidebar extends React.Component {
-  getSidebarContent = () => {
+  render() {
     const {
-      location,
-      navigation,
-      children,
       constants,
+      navigation,
+      location,
+      children,
       parentPath
     } = this.props;
-    return (
-      <NavigationAccordion
-        constants={constants}
-        navigation={navigation.navTabs}
-        location={location}
-        parentPage={parentPath}
-      >
-        {children}
-      </NavigationAccordion>
-    );
-  };
-
-  render() {
-    const { constants, navigation } = this.props;
     const { SITE, BASEURL } = constants;
-    const { title, tag, path } = navigation;
+    const { title, tag, path, navTabs } = navigation;
+    // check if the site only has one page (one navTab with no sub pages)
+    const isSinglePageSite =
+      navigation &&
+      navTabs &&
+      navTabs.length < 2 &&
+      (!navTabs[0].pages || navTabs[0].pages.length === 0);
 
     return (
       <div
@@ -36,14 +29,29 @@ export default class Sidebar extends React.Component {
         className="sticky-mm scroll-auto-mm scroll-styled viewport-almost-mm px12-mm"
         style={{ top: '10px' }}
       >
-        <div className="my12 border-b border--darken10 pb12">
+        <div className="mb12 border-b border--darken10 pb12">
           <ProductMenu
             productName={title || SITE}
             tag={tag || undefined}
             homePage={`${BASEURL}/${path || ''}`}
           />
         </div>
-        {this.getSidebarContent()}
+        <div
+          className={classnames('', {
+            // if there is only one navTab (with no sub pages)
+            // hide the NavigationAccordion on mobile
+            'none block-mm': isSinglePageSite
+          })}
+        >
+          <NavigationAccordion
+            constants={constants}
+            navigation={navTabs}
+            location={location}
+            parentPage={parentPath}
+          >
+            {children}
+          </NavigationAccordion>
+        </div>
       </div>
     );
   }
@@ -60,15 +68,10 @@ Sidebar.propTypes = {
   }).isRequired,
   frontMatter: PropTypes.shape({
     navOrder: PropTypes.number,
-    headings: PropTypes.array,
-    sidebarTitle: PropTypes.string
+    headings: PropTypes.array
   }).isRequired,
   parentPath: PropTypes.string,
   children: PropTypes.node,
-  layoutConfig: PropTypes.shape({
-    hideSubItems: PropTypes.bool,
-    includeFilterBar: PropTypes.bool
-  }),
   constants: PropTypes.shape({
     SITE: PropTypes.string.isRequired,
     BASEURL: PropTypes.string.isRequired