+++ src/components/page-layout/examples/example.js
@@ -27,7 +27,7 @@ export default class Basic extends React.Component {
           }
         }}
         location={{
-          pathname: '/PageLayout/'
+          pathname: '/PageLayout/examples/'
         }}
         frontMatter={{
           title: 'Examples',
@@ -38,12 +38,11 @@ export default class Basic extends React.Component {
         AppropriateImage={AppropriateImage}
         navigation={{
           hierarchy: {
-            '/PageLayout/': {
-              parent: '/PageLayout/',
-              title: 'Demo'
+            '/PageLayout/examples/': {
+              parent: '/PageLayout/examples/',
+              title: 'Examples'
             }
           },
-
           navTabs: [
             {
               path: '/PageLayout/',
@@ -81,8 +80,8 @@ export default class Basic extends React.Component {
             }
           ]
         }}
-        topics={{
-          '/PageLayout/': {
+        filters={{
+          '/PageLayout/examples/': {
             contentType: 'example',
             description: 'Replace me.',
             language: ['JavaScript'],
@@ -90,68 +89,43 @@ export default class Basic extends React.Component {
             navOrder: 3,
             path: '/PageLayout/',
             title: 'Examples',
-            topics: [
+            topics: ['Cool', 'Awesome'],
+            pages: [
               {
-                count: 3,
-                name: 'Cool',
-                pages: [
-                  {
-                    contentType: 'example',
-                    description: 'Replace me.',
-                    language: ['JavaScript'],
-                    layout: 'example',
-                    path: '/PageLayout/example-1/',
-                    text: 'Example 1',
-                    thumbnail: 'header-image',
-                    title: 'Example 1',
-                    topic: ['Cool'],
-                    url: '/PageLayout/example-1/'
-                  },
-                  {
-                    contentType: 'example',
-                    description: 'Replace me.',
-                    language: ['JavaScript'],
-                    layout: 'example',
-                    path: '/PageLayout/example-2/',
-                    text: 'Example 2',
-                    thumbnail: 'header-image',
-                    title: 'Example 2',
-                    topic: ['Cool'],
-                    url: '/PageLayout/example-2/'
-                  },
-                  {
-                    contentType: 'example',
-                    description: 'Replace me.',
-                    language: ['JavaScript'],
-                    layout: 'example',
-                    path: '/PageLayout/example-3/',
-                    text: 'Example 3',
-                    thumbnail: 'header-image',
-                    title: 'Example 3',
-                    topic: ['Cool', 'Awesome'],
-                    url: '/PageLayout/example-3/'
-                  }
-                ],
-                url: '/PageLayout/#cool'
+                contentType: 'example',
+                description: 'Replace me.',
+                language: ['JavaScript'],
+                layout: 'example',
+                path: '/PageLayout/example-1/',
+                text: 'Example 1',
+                thumbnail: 'header-image',
+                title: 'Example 1',
+                topics: ['Cool'],
+                url: '/PageLayout/example-1/'
               },
               {
-                count: 1,
-                name: 'Awesome',
-                pages: [
-                  {
-                    contentType: 'example',
-                    description: 'Replace me.',
-                    language: ['JavaScript'],
-                    layout: 'example',
-                    path: '/PageLayout/example-3/',
-                    text: 'Example 3',
-                    thumbnail: '/files/ios-horizontal.png',
-                    title: 'Example 3',
-                    topic: ['Cool', 'Awesome'],
-                    url: '/PageLayout/example-3/'
-                  }
-                ],
-                url: '/PageLayout/#awesome'
+                contentType: 'example',
+                description: 'Replace me.',
+                language: ['JavaScript'],
+                layout: 'example',
+                path: '/PageLayout/example-2/',
+                text: 'Example 2',
+                thumbnail: 'header-image',
+                title: 'Example 2',
+                topics: ['Cool'],
+                url: '/PageLayout/example-2/'
+              },
+              {
+                contentType: 'example',
+                description: 'Replace me.',
+                language: ['JavaScript'],
+                layout: 'example',
+                path: '/PageLayout/example-3/',
+                text: 'Example 3',
+                thumbnail: 'header-image',
+                title: 'Example 3',
+                topics: ['Cool', 'Awesome'],
+                url: '/PageLayout/example-3/'
               }
             ]
           }